refactor: parsing 로직 view로 이동
