+++ src/main/java/lotto/view/InputView.java
@@ -1,6 +1,8 @@
 package lotto.view;
 
+import java.util.List;
 import java.util.Scanner;
+import java.util.stream.Collectors;
 
 public class InputView {
 
@@ -27,6 +29,26 @@ public class InputView {
         clearBuffer();
         System.out.println("지난 주 당첨 번호를 입력해 주세요.");
         return SCANNER.nextLine();
+    }
+
+    public List<Integer> getWinnerLottoNumbers(String lottoNumbers){
+        List<String> lotto = parseWinnerLotto(lottoNumbers);
+        return lotto.stream()
+                .map(this::parseNumber)
+                .collect(Collectors.toList());
+    }
+
+    private int parseNumber(String lottoNumber) {
+        try {
+            return Integer.parseInt(lottoNumber);
+        } catch (NumberFormatException ex) {
+            throw new IllegalStateException("숫자 형식이 올바르지 않습니다.");
+        }
+    }
+
+    private List<String> parseWinnerLotto(String lastWinnerLotto) {
+        String text = lastWinnerLotto.replaceAll(" ", "");
+        return List.of(text.split(","));
     }
 
     private void clearBuffer() {
